# api-gateway/pkg/response/response.go
package response

import (
	"github.com/gofiber/fiber/v2"
)

type APIResponse struct {
	Success bool   `json:"success"`
	Message string `json:"message,omitempty"`
	Data    any    `json:"data,omitempty"`
	Error   string `json:"error,omitempty"`
}

func Success(c *fiber.Ctx, data any, message string) error {
	return c.JSON(APIResponse{
		Success: true,
		Message: message,
		Data:    data,
	})
}

func Error(c *fiber.Ctx, statusCode int, err error) error {
	return c.Status(statusCode).JSON(APIResponse{
		Success: false,
		Error:   err.Error(),
	})
}